feat(#86): refactor: context-info.ts — audit and modularize
